Explicit (uint32_t)serialize_size()
The cast is safe because at the end of serialize_size()
[kvlds/serialize.c line 444], we have:

    /* Sanity check: we can't store more than this in the node. */
    assert(size <= UINT32_MAX);

At the top of serialize_size(), we have a return(N->pagesize), but
->pagesize is already a uint32_t, so that's also safe.

+++ kvlds/btree.c
@@ -288,7 +288,7 @@ btree_init(struct wire_requestqueue * Q_lbs, uint64_t npages,
 	if (T->root_dirty != NULL) {
 		/* Record the size of the serialized node. */
 		T->root_dirty->pagesize =
-		    serialize_size(T->root_dirty);
+		    (uint32_t)serialize_size(T->root_dirty);
 
 		/* Figure out the oldestleaf. */
 		if (T->root_dirty->type == NODE_TYPE_PARENT) {
